Update Makefile with generic compilation rule
Added a generic way to compile C source files in the Makefile. This change allows any C file to be compiled into an executable by simply using its filename (without the .c extension) as the make command argument. The goal is to speed up the workflow and reduce redundancy in our Makefile.