Add new program to calculate array sizes
This commit adds a new program file, `12.c`, in the `base` directory. The program is designed to calculate the sizes of integer and character arrays, as well as the number of elements in these arrays. This can be helpful for understanding memory usage in different parts of the program or system. More detailed commentary has been included in the source code for better understanding of the calculations made.